Tweaked ci configuration for more consistency (#60)

+++ tests/env.run.sh
@@ -26,7 +26,15 @@ echo " == Jahia became alive in ${ELAPSED_TIME} seconds"
 
 # Add the credentials to a temporary manifest for downloading files
 mkdir /tmp/run-artifacts
-cp ${MANIFEST} /tmp/run-artifacts
+# Execute jobs listed in the manifest
+# If the file doesn't exist, we assume it is a URL and we download it locally
+if [[ -e ${MANIFEST} ]]; then
+  cp ${MANIFEST} /tmp/run-artifacts
+else
+  echo "Downloading: ${MANIFEST}"
+  curl ${MANIFEST} --output /tmp/run-artifacts/curl-manifest
+  MANIFEST="curl-manifest"
+fi
 sed -i -e "s/NEXUS_USERNAME/${NEXUS_USERNAME}/g" /tmp/run-artifacts/${MANIFEST}
 sed -i -e "s/NEXUS_PASSWORD/${NEXUS_PASSWORD}/g" /tmp/run-artifacts/${MANIFEST}
 
@@ -36,5 +44,15 @@ echo " == Warming up the environement =="
 
 echo " == Environment warmup complete =="
 
+mkdir /tmp/results/reports
+
 echo "== Run tests =="
 yarn e2e:ci
+if [[ $? -eq 0 ]]; then
+  echo "success" > /tmp/results/test_success
+  exit 0
+else
+  echo "failure" > /tmp/results/test_failure
+  exit 1
+fi
+# After the test ran, we're dropping a marker file to indicate if the test failed or succeeded based on the script test command exit code
